requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
python-dateutil==2.8.2
pytz==2023.3
sqlparse==0.4.4
asgiref==3.7.2
typing-extensions==4.8.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
urllib3==2.0.7